add current date as a balance data-point
this has bugged me for a while. 
I want to see when my account hasn't changed. To that end I want another data-point for the current day I'm on (even though it's not a transaction). This will help complete my picture of activity/ inactivity on my account. 

The line I added might not add the new entry to quite the right place in the array ( I note transactions_with_totals is doing some sorting and reversing D: ), so this def needs a smoke-test